making windows update system, changed patch file

+++ Client/updater/updater.cpp
@@ -65,11 +65,11 @@ bool checkForUpdatesWindows(std::string exec_path) {
     while(exec_path.back() != '/')
         exec_path.pop_back();
 
-    /*if(std::filesystem::exists(sago::getDataHome() + "/Terralistic/update.patch")) {
+    if(std::filesystem::exists(sago::getDataHome() + "/Terralistic/update.patch")) {
         std::system(((std::string)"cd \"" + exec_path + "\" && patch -t -p3 < \"" + sago::getDataHome() + "/Terralistic/update.patch\"").c_str());
         std::filesystem::remove(sago::getDataHome() + "/Terralistic/update.patch");
         return true;
-    }*/
+    }
 
     return false;
 }
